pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["Akshay Prabhu <myemail@email.com>"]
readme = "README.md"
packages = [{ include = "django_template" }]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-environ = "^0.11.2"
whitenoise = "^6.5.0"
django-simple-history = "^3.4.0"
django-crispy-forms = "^2.0"
crispy-tailwind = "^0.5.0"
setuptools = "^68.2.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
ruff = "^0.0.290"
django-browser-reload = "^1.11.0"
pip = "^23.2.1"
install = "^1.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
